requirements.txt
langchain==0.3.19
faiss-cpu==1.10.0
hydra-core==1.3.2   
fastapi==0.118.0
uvicorn==0.37.0
openai==1.61.1
langchain-community==0.3.17
langchain_openai==0.3.6
azure-core==1.35.1         
azure-functions==1.23.0                 
azure-identity==1.25.0             
azure-storage-blob==12.26.0    